Create illustration placeholders
Gives the "image" selector to divs used as the illustration placeholder.
The image selector is styled to have a blue border with rounded corners.
Also, the descendant combinator is used on the image in the first body
so that it keeps its grey background instead of having a blue border and
white background.