Change global font and create quote classes
The font for all text on the page is changed to roboto using Google's font hosting service via a link  in the html head.

The quote and quote author have been given the classes "quote" and "author", respectively.

Also, the background color of body three is changed to a grey color.